Make ModuleInfo derive a bunch more traits
Summary: I want to be able to store a ModuleInfo in an Error, since we need the module contents and line index to print source contexts. But Error derives Hash, Ord, PartialOrd, Eq, and PartialEq. This diff makes ModuleInfo use ArcId so it can derive Hash, Eq, and PartialEq. (The next diff removes Ord and PartialOrd from Error, so we don't need to derive those.)

Reviewed By: yangdanny97

Differential Revision: D75571870

fbshipit-source-id: 8f2366a3dbb90a0440c06c7ef653bbe3b60e3a6c