Always lookup hashed names
Summary:
I need to add the ability to track usage in name lookups, which is going
to lead to the need for yet another case where we have an implementation and
then shims (one for usage-tracked lookups and one for other lookups).

This is annoying but fine, the difference between usage-tracked lookups and
non-usage-tracked lookups is fundamentat.

But if we also have hashed and unhashed versions, now we wind up with 4
variations which becomes a bit overwhelming, and the difference between hashed
and unhashed lookups isn't real - it's super easy to just hash on the caller
side (more than half of the lookups today are hashed already anyway)!

So I'm just deleting the existing `lookup_name` and renaming `lookup_name_hashed`
to `lookup_name`.

Reviewed By: grievejia

Differential Revision: D75801876

fbshipit-source-id: e0a8d6d22df4e5980b6e8ef7dcf53d334e84d0ac