Implement usage tracking for AugAssign
Summary:
This is slightly messy, since `Binding::AugAssign` takes the entire
statement as an input which isn't what the `target.rs` code is optimized around.

The solution I came up with (clone and replace the value) costs one extra
expression clone; we could potentially optimize it later, but I doubt that augmented
assigns with really huge RHS expressions are common enough that this matters for
practical purposes.

Reviewed By: grievejia

Differential Revision: D75749459

fbshipit-source-id: b668b607ba789e537a4611d0f76934f98a74a45e