Thread usage into lookup name logic.
Summary:
This is finally the end of large-scale refactors of `bindings` code:
everything is now in place to start actually working on recording
and powering type inference with usage information.

From this point forward, we'll be working on the trickier but much
smaller-scope (in terms of amount of code it touches) problem of
how and when to actually record usage, and what to do downstream
in the solver.

Reviewed By: grievejia

Differential Revision: D75801877

fbshipit-source-id: 6e3bf10f1db10c4e9735818dd8c8f776505f11ff