Rework the class object vs ClassIndices code a bit
Summary:
I realized that putting the class object idx inside `ClassIndices` isn't really
optimal after all; we get some nice benefits from the lifetime semantics of
`User`, which (correctly) wouldn't like us leaking the user into the `Scopes`
object.

So I split it back off - it gets created at the same time, but it only lives for
the lifetime of creating a class def and gets consumed when we bind the
class object to the class name, which is exactly what should happen.

This prepares class.rs for usage tracking, which I'll do in the next diff;
keeping this diff as small as possible since it's slightly more involved than
I want the actual cutover diffs to be.

Reviewed By: grievejia

Differential Revision: D75743592

fbshipit-source-id: 365eb178cac7515205f89d652772094644a4d2c0