Introduce UsageLink, convert `expr.rs` to usage tracking
Summary:
Finish converting `expr.rs` to usage tracking; it already had it just by passing
the `usage` down except for yields, which is what we handle here.

With yields, we encounter a complication in the usage tracking:
- We absolutely do need yields to be included in first-usage tracking, or else
  any yield (which could be not just a simple yield but a complex expression
  involving function calls, etc) would get skipped in our usage tracking logic.
- As a shortcut, we *could* use the enclosing usage (the yield appears inside some
  larger expression - possibly as the entire expression) and we *do* have a Usage
  for that. But this would be sub-optimal because in some cases the yield itself
  might tell us something about the type, for example if we `yield x[0]` from an
  empty list in a function marked as returning a `Generator[int, None, None]`, we
  can infer `int` as the type of the list but we would get `Any` if we did it
  the lazy way.

The solution I came up with - partially motivated by the fact that we are going to
run into the same issue for a couple of `BindingExpect` cases when we convert
`stmt.rs` to statement tracking shortly - is to just use a placeholder `Key/Binding`
pair that points at the `KeyYield` / `KeyYieldFrom`; this placeholder will be our
"user", that way when we implement first-usage type inference it will force a calculation
of the `BindingYield` / `BindingYieldFrom` which will give us a chance to pin `Var`s
with contextual typing.

An alternative would be to not force `Usage` to be based on `Key` and insted let it
be based on any `K: Keyed`, but I'm convinced this is much better:
- It makes the code much simpler for the vast majority of use cases - there is only
  a tiny handful of cases where we need some other variety of binding to be
  usage-tracked
- It also almost certainly saves us RAM; at the cost of one extra key/binding pair
  in the relatively few use cases where this pops up, we get the ability to use a
  bare `Idx<Key>` with no tag when we start actually tracking usage - which will
  wind up affecting the size of *every* `Binding`. The alternative would probably
  require an enum type, and that would cost RAM in more places.

Reviewed By: grievejia

Differential Revision: D75744942

fbshipit-source-id: f8a9b0a39388fc139bb8d935f8e567beed44047d