Convert `Delete` handling to usage tracking
Summary:
This requires another UsageLink so that we can tie first-usage
to a `KeyExpect` node. It is important that we not skip it because in principle
the `del` target could be anything including a complex expression that would
pin the type of some `Var` (it's not that unusual to `del` a dict key, for example,
to save RAM in a space-intensive Python program).

Reviewed By: grievejia

Differential Revision: D75747890

fbshipit-source-id: 4cbc839d468d12c6b76c2109caba5272df1f409a