Simplify `lookup_name_impl`
Summary:
It doesn't need to be so complicated - there's no reason to be handling
the `barrier` case in the flow-sensitive part of the loop, it gets handled
by the static part.

That means the flow-sensitive logic becomes invariant over normal vs
mutable lookups (which is reasonable - it's always fine to mutate a local, so
there's no reason we ought to care).

There's no behavior change here, just more clarity and less code.

Reviewed By: grievejia

Differential Revision: D75879419

fbshipit-source-id: e9897672d305d3e675452037a75292ee8416cb3e