Add more detail to `iterate` doc comment
Summary:
It's a bit weird that we are returning a `Vec` here. What does it
represent? The in-order-types of a tuple? That would be totally plausible...

No! It is an implementatin punting union handling to callers. Better to say
so explicitly, because nothing in the function name or type signature makes
it obvious what the `Vec` is doing here.

Reviewed By: grievejia

Differential Revision: D75883602

fbshipit-source-id: 8599d34d3b1b6fa72e89a004a7bc385b79baf594