Rename `b` to `to_unpack` in `UnpackedValue`
Summary:
Given that this is just an `Idx<Key>`, I find that a name telling me more
about what it represents helps with readability. This is the binding of
the thing we need to unpack (so for example in `x, y = foo()` it's the
binding for `foo()`).

Let's call it that.

Reviewed By: grievejia

Differential Revision: D75883138

fbshipit-source-id: ba1e1fe3ee2c081f93cb2ed5320c4d74df437ecc