Make configs use kebab-case by default, but keep snake case for existing options
Summary:
At PyCon, it seemed like people expected the config keys to match flag names in being kebab-case. It's something I've been considering changing since before PyCon, but this is a good motivation to do it.

This diff renames all flags to be kebab-case, but also lets existing, multi-word flags accept the snake_case form to be backward compatible. Our config documentation is also updated to reflect this, and I add a test in `config.rs` to make sure nothing new breaks.

Reviewed By: ndmitchell

Differential Revision: D75892768

fbshipit-source-id: e821c57838753e98d95484b796dbf4539cf60933