fix comments for landing page animation delay
Summary:
We added the landing page animation before pycon and made a couple different changes, resulting in the current set of comments being out of date. Updating the comments in this diff.

A better solution here is likely to create a file that shows the ordering of the elements to animate in the form of a list, and then associate a delay with each element. However, this is a bit harder to do for files using stylex to set the delay, since it requires a string instead of an int. I don't think this is worth doing right now, but if we find ourselves changing the delay often in the future, which I really don't see happening, then we can make that change

Reviewed By: ndmitchell

Differential Revision: D75962104

fbshipit-source-id: 1b1cf920bcf0ca0cca10834b75b3c2cc9c3e06ee