Expose answers to variance algorithm
Summary:
The current variance bindng structure is wrong.

The reason is that we not only need the class's fields, methods and base classes. We need the entire class hierchy so that we have the ability to lookup the base class's fields and methods etc.

To get all the base classes from a class:&Class, I am experimenting doing something like:

        `let metadata = get_metadata(class);
        let base_types = metadata.bases_with_metadata();`

(where we get the metadata closure from the solver)

I think this means we no longer need base_types in our binding.

I would also like to add the ability to lookup a class based on it's name, so I'm looking into something like:

` let metadata = solver.get_metadata_for_class(class);

                let ancestor_class = metadata.ancestors(solver.stdlib).find(|ancestor| {
                    let class_obj = ancestor.class_object();
                    class_obj.name() == class_name
                });

                let my_class = match ancestor_class {
                    Some(ancestor) => ancestor.class_object(),
                    None => class,
                };`

but I'm not sure yet if this will fully work.

What I do know is that we should probably expose the solver to the inference algorithm.

Next steps:

- verify that the way I get the base types and obtain a class from a string is actually right, and then remove the base types from the binding.

- figure out what to do about fields. Again, what we have now is not enough, so ideally we can lookup the fields with a query containing the class name or the &class object.

Reviewed By: rchen152

Differential Revision: D75887001

fbshipit-source-id: 1761a8105dbbfdf7b1d9b86f1717f48547e81440