write `get_possible_attribute_bases` function
Summary:
We will use this helper to split up the attribute access handling. First we get a list of all the bases, considering unions, bounded type vars, etc. Then we try to do the accesses and process the results.

This handles a lot more cases and has better behavior than the existing logic, but the control flow is nested quite deeply so LMK if anyone has suggestions to improve it

This will be used in the next diff

Reviewed By: stroxler

Differential Revision: D75878379

fbshipit-source-id: 1c127931cfaaef78fc0afd3ebe172c4746f4e737