Return completion item and hover with symbol information
Summary:
In this diff, we start to use the binding table to generate a symbol kind for various IDE services with best effort. Notably, we will treat all imported symbols as `SymbolKind::Variable` for now.

In this diff, I connected it to hover and completion. Later the same information can be used to implement semantic tokens.

Reviewed By: SamChou19815

Differential Revision: D75711009

fbshipit-source-id: 9450ff2c589b264669782b321bd1b5dc376b5529