Add more tests where deferred inference matters
Summary:
Deferred inference is not just for empty container literals, we also
use it whenever a generic function (most often, but not necessarily,
a constructor) fails to fully determine the result type; that means
we get OCaml "weak-type-variable" like inference on user-defined
container-like types and on functions that produce empty containers
just like we do on literals like `[]` and `{}`.

Adding some tests to demonstrate this.

Reviewed By: ndmitchell

Differential Revision: D75960137

fbshipit-source-id: 3f572c865567409db26e9040b0e3ceda830ac905