Tweak misleading error messages about `__setitem__` and `__delitem__`
Summary: We had error messages that said item assignment and deletion were "not supported," which is confusing when the corresponding magic method does exist but was called with the wrong arguments. Change these messages to be more generic. (These messages always appear as the first line of a multi-line message; the rest of the message provides more details.)

Reviewed By: stroxler

Differential Revision: D76011309

fbshipit-source-id: 965ddcc6125c77b5aeb9189691ca888001ed52d3