Make legacy typevars with unspecified variance invariant
Summary:
I noticed that when collecting variance information for various predefined baseclasses like `typing.Collection[Variable[typing._KT];` etc.. the variance is set to "undefined".

It is unclear to me if this overrides the `infer_variance` case, so if so, this is something I can handle later, but for now, the variance here is wrong everywhere since I believe legacy params with unspecified variance should be invariant, and this should happen before we even begin the algorithm, since later, there is no distinction between legacy vs scoped typed vars.

Reviewed By: rchen152

Differential Revision: D75919590

fbshipit-source-id: fd5f631c6e233c34abc35e1074140f2c9431a054