Split the `AnnAssign` logic by value or no value
Summary:
I want it to be easier to share logic between `AnnAssign` and the
single-target case of `Assign`, because they both will need to handle
first-use-pinning of placeholder types (note: it's easy to imagine that
`AnnAssign` wouldn't need placeholders, but the annotation might be
just a qualifier).

That requires splitting the logic so that the yes-value / no-value
cases are handled separately.

The resulting logic is a bit more verbose, but also limits the scope
of local variables better so it's arguably simpler. More importantly,
it opens the door do further refactors.

Reviewed By: rchen152

Differential Revision: D76050906

fbshipit-source-id: bca25769f4c1d6838e4b46e6985c4915f10d99f9