Factor AnnAssign annotation binding out
Summary:
I have been having trouble reading the code for `AnnAssign` (which is
very relevant to first usage because - due to the possibility that
the annotation is just a qualifier and not a type - it's one of the cases
where we want to allow unpinned placeholder types that get pinned on first
usage).

It's not so much that the code is confusing, it's just too big a block to
skim, with too many local variables in scope at the same time for my taste.

Factoring the annotation binding - which is already pretty well separated
from the rest of the code - helps.

Reviewed By: rchen152

Differential Revision: D76050908

fbshipit-source-id: e6587eb4dc3bfbb94c69ba6a828924019a36163f