Add in a couple comments to the AnnAssign logic
Summary:
I thought a couple of hooks in the code were not particularly
self-explanatory and deserved an inline comment explaining
what's going on. I particularly dislike it when we are special-casing
part of the standard library and it's not made clear in a comment,
someone onboarding could waste significant time trying to chase that
down.

Reviewed By: rchen152

Differential Revision: D76050907

fbshipit-source-id: fed842c2f1eff6b5bd3a14c45d6e541adb360027