Add ability for --count-errors to count all errors
Summary:
The help text for --count-errors says that there's a way to get counts of all errors, not just the top N, but this doesn't seem to be implemented. Seems like a useful feature, so this diff adds it.

I also flipped the order that error counts are printed from least to most, to most to least, for consistency with --summarize-errors.

Reviewed By: ndmitchell

Differential Revision: D76104864

fbshipit-source-id: 97ee463b0775137f5c1c376dc98cad457558d075