create delayed component
Summary:
## This stack
Refactoring the way that we are doing animation delays. This is currently set by using an integer to define delays and it's scattered across multiple files. We also rely on a decently sized `useEffect` block in each component to set this. To make this better, I am going to:
1) Create a wrapper component to control animation delay
2) Centralize animation delay in one file so that it's easier to visualize all the components that are animating, and seeing the order in which they get animations

## This diff
Starting off by creating a delayed component, this will be used in the following diff: D75992218

Reviewed By: connernilsen

Differential Revision: D75988581

fbshipit-source-id: 1da942ff9a0a9ec2ea2b180c3fd37a0ae6e55e18