Introduce a function that locate covering AST node by location, along with all the parent nodes
Summary:
This diff introduces an AST utility which, given a module and a text location, returns all the AST nodes that "cover" the location.

In practice, there's usually more than 1 covering node for any location. E.g. the `ModModule` itself always cover any location (as long as the location is within range). Even for simple code like this:
```
def foo(x):
  return x + 1
#        ^ location
```
The function def of `foo`, the compound expression `x+1`, and the name expression `x`, all "cover" the location. The way we handle this is to have our function return a list of AST nodes, sorting by how tightly they cover the location. In this case, this first node in the list will be name `x` itself, followed by `x+1`, followed by the containing function `foo`, followed by the entire module.

This utility function will be used in subsequent diffs where we are going to enhance the capability of identifier lookup in our language server.

Reviewed By: ndmitchell

Differential Revision: D76174531

fbshipit-source-id: 1dba2bdfd8e158923a66902fe177dda98f4144ae