Use covering node to compute `identifier_at`
Summary:
This diff uses the newly introduced `locate_node` function from D76174531 to implement `identifier_at()`.

Semantically speaking, nothing really changed in this diff -- both our own visitor and `locate_node()` essentially looked for the same thing. The point of this change is to prepare the code so we could extend `identifier_at()` later to find more nodes and include more context, which is not possible with the pre-existing logic.

Reviewed By: ndmitchell

Differential Revision: D76174530

fbshipit-source-id: abad6f60d64f7df1ba41fa46d511a69701c1f371