Create helper to calculate animation delay to remove hardcoded delays
Summary:
This diff moves the calculation for the animation delay into a single file: `componentAnimationDelay.tsx`, rather than defining it manually in components scattered across the codebase. We define an ordered list of components that needs to be animated, along with the delay that should be set between each element, and then dynamically calculate the delay for each component.

This gives a slightly better feel to the animation delay since the time between each component is now uniform, along with making it easier to make changes in the future for the loading animation of components.

Reviewed By: stroxler

Differential Revision: D76064615

fbshipit-source-id: 0d80af16ad3cdb3b807f3b94578727368524aa98