Initial auto import support via code action
Summary:
This diff implements an MVP support for auto import, via code action. We start from an unknown-name error, and try to create a code action that fixes it by adding an import statement. To find the relevant import, we will search all the export table (but notably excluding those exports like `b` in `from a import b`).

In the future, we should really have an indexing setup so we can have fuzzy search to power it for autocomplete as well.

Reviewed By: kinto0

Differential Revision: D76222223

fbshipit-source-id: 0b22667be3152f8edf0ff1b88ad579114d01d404