Add module import context to `identifier_at`
Summary:
This diff builds on top of D76174532 and adds the first kind of context: imports. Concretely, we return `IdentifierContext::ImportedModule` when we detect that that identifier the user looks for appears as the name of the imported module (i.e. `x.y` in `import x.y` or `from x.y import ...`).

This serves as a nice testbed to give us a taste of what the new `identifier_at()` API is capable of doing. Although we do have import handling today, the pre-existing logic was not able to depend on `identifier_at()` and had to do its own AST walk precisely due to the lack of context in the old version of `identifier_at()`. With contextual info, we are able to replace the pre-existing import look up logic with `identifeir_at()` without having a custom version of it just to search for import names.

Reviewed By: ndmitchell

Differential Revision: D76233356

fbshipit-source-id: 75680e9bd6d4f8481eb55a46d113cf7eff22f7be