Include context in the return type of `identifier_at`
Summary:
One of the issue with the `identifier_at()` function is that it can only tell its callers which identifier the user is clicking, without any surrounding context. This means that we'll have no idea if the user clicks on a name of func definition, a name of keyword argument, a class name of a match branch, or just the name of a local variable.

With D76174531 and D76174530 introducing a way to compute covering nodes, we can now extend `idenfier_at` to address the aforementioned limitation. This diff starts to change the return type of that function to a pair of identifier and its surrounding context. We are not adding any functionality right now, which means that we only have 1 kind of context at the moment: the identifier appears as part of an expression. But we'll add more of them in subsequent diffs with the goal to make gotodef smarter in how it determines the target.

Reviewed By: ndmitchell

Differential Revision: D76174532

fbshipit-source-id: 995b2e270c1c6873a4e7eee943081894d2b82dcc