Remove output_path in config_migration
Summary:
MaggieMoss shared that members on the team has found it confusing to use an `output_path` in the `pyrefly init` flow. This was already removed in D74816787, but there's some code in `config_migration.rs` which still references `output_path`. This diff cleans this up. The logic now simply becomes:

If the input path ends with `pyproject.toml`, then `pyproject.toml` will be used as the `output_path`. Otherwise, use `Pyrefly.toml`

Reviewed By: MaggieMoss

Differential Revision: D76288613

fbshipit-source-id: ae5fc36d30ed1b1ce849eb431635ec89a51f920b