Support signature help
Summary: This diff implements signature help. We can generate signature help by inspecting the type of the callee when it can be converted into a `Type::Callable`. In addition, for overloads, we will use the information recorded in the last diff to show the signature help of the likely chosen overload, as well as all other options.

Reviewed By: kinto0

Differential Revision: D76222222

fbshipit-source-id: 51569c32fb9450faa3721d3d89d1024da4f2be7a