Rename `bind_assign_impl` to `bind_target_name`
Summary:
The new name better represents what this is: it's the leaf node
of the `bind_target_impl` recursion tree that we hit when we
find an `ExprName` target.

It does not actually create a `NameAssign` node, nor is it
reliably used in assignment statements (there are some cases
where it is), so the old name could be a little confusing.

Reviewed By: yangdanny97

Differential Revision: D76201678

fbshipit-source-id: 50a24e8ee65242d54476d0aa9c763d94a4302e5a