Track missing pattern-match guard as a bug
Summary:
For fixing Var-nondeterminism, I need to introduce a specific discipline around how
usages work for narrowing. Before I do that, I think it makes sense to introduce narrowing
in the one place where I know for sure it isn't working properly that would directly
affect the approach I'm going to take for determinism.

This diff adds a test case of match guard narrowing, which we don't yet implement. I'll
fix it in the next.

Reviewed By: yangdanny97

Differential Revision: D76201676

fbshipit-source-id: f9483626dcc6e1b3b090ae0b1903379847fa16c1