Put a comment in about the alias special case
Summary:
I don't think the function names are enough to describe what's going on
here, since "alias" seems like it might mean TypeAlias or something like
that.

It seems better to have a comment state in plain English that this applies
only to stdlib stubs, and we're talking about the `_Alias()` construct.

Reviewed By: yangdanny97

Differential Revision: D76199674

fbshipit-source-id: b9db7bcd1c19f1a33f2f2be6beabf09f49557a56