Update python interpreter in config to make verbose logging clearer (#450)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/450

It's confusing that we say `python_interpreter: None` in the `DEBUG Config is:` output during verbose logging. It's confused me and several users, so let's fix that bug here so that we can stop being confusing.

p.s. I think I made this change a while ago and just forgot to write the found interpreter :(

Reviewed By: yangdanny97

Differential Revision: D76307599

fbshipit-source-id: a0e6cc54bc996d5f85aed798e812c63acdd7e5da