Eliminate expr duplication for multi-target assign [2/2]
Summary:
I realized that, in addition to the contextual typing soundness bug tracked in
the previous diff, the use of a single `Expr` for multiple targets will be
able to produce a `Var` leak if the assignment happens to be a first usage,
creating a hole in first-usage-based nondeterminism.

Given that there's both a determinims issue *and* the contextual typing is a
static analysis bug in the first place, I'm convinced that the correct thing to
do is disable contextual typing for multi-target assignment.

Note that I special case a single target to do what we did before, because
this code path is used in many contexts where there could be just one target
(for example, a single assignment to an attribute or a subscript). So it would
be an unnecessary an nontrivial regression to turn of contextual typing in that
case.

The unit test makes the error look ambiguous because it doesn't say that `x`
is the problem (as opposed to `y`) but in practice this isn't a huge problem;
the range on the error is the target, so the error will be attached to `x`
rather than `y`. Maybe it would be better to pass the target through so we could
use it in the error message, but that costs additional clones and boxes; I'm not
sure it's worth it given how rarely multi-target assign is used anyway.

Reviewed By: yangdanny97

Differential Revision: D76267529

fbshipit-source-id: 18d6904647aaf3aaad864a67777423cb7d66a4d7