Introduce and use `Usage::Narrowing`
Summary:
I have concluded after some careful thought that the only short
term path forward I see to avoid any chance of nondeterminism is
if narrowing operations cannot pin placeholder types, and instead
we always force placeholders before narrowing.

It's quite difficult to not do this because in narrowing, we duplicate
the same expression (with the same lookups) potentially many times. If
the narrowing test is a first-usage, that means there are an unlimited
number of bindings now exposed to the placeholder types, and unless
we can force a deterministic ordering - which seems difficult, the
ways of I can think of introduce recursion and are far worse than
this approach - it seems very difficult to avoid leaking the placeholders
in ways that could hit race conditions.

I'm pretty sure it would be very rare for this to cause actual loss of type
inference for users anyway because a narrowing test is very rararely going to
cause a placeholder type to be inferred anyway; I think a generic type guard is
the main case I can see, and it probably would be a little confusing to pin in
that case anyway.

----

NOTE: I'm pretty sure in the long run we need some discipline - maybe a rust
wrapper type, maybe something else - to ensure that a narrowing test cannot
use an improperly ensured expression. I've audited all narrows manually for
this diff, but without an abstraction I think it's extremely likely that
we accidentally introduce a race condition through an incorrect ensure later.
I'll try to follow-up on this, but it can wait until we actually implement
var-pinning.

Reviewed By: grievejia

Differential Revision: D76267532

fbshipit-source-id: 4a56e3fea971183467b487935928d42e3f6084a4