Simplify if statement narrowing a bit
Summary:
I'm about to make narrowing never act as a usage that could pin placeholder
types, because (a) narrowing almost never would pin them anyway and (b) it
is very difficult to avoid nondeterminism unless we make this change, given
that narrowing causes the same expression to fan into multiple bindings.

With that change coming up, the If handling can be simplified, which also
makes it a lot clearer what's really going on (it's pretty strange otherwise
to see a `NarrowOps::from_expr` on something that looks as if it might
never be ensured... the types and control flow make it such that this is
okay, but you really have to squint to see it!).

Since the bigger narrowing-usage change is a rather large but simple refactor,
I'm tackling trickier local changes like this one first.

Reviewed By: grievejia

Differential Revision: D76267531

fbshipit-source-id: 2edb17da656b0d8bd6cf919b4a6732e404d09943