Add class/function definition context to `identifier_at()`
Summary:
This diff enhances `identifier_at()` and teach it to recognize cases where the user cmd+click on the function/class name in the `def` or `class` statement directly.

Previously, we were relying whether definition_at() can find the definition key of the function/class in the scope stack, which requires a walk of the entire scope. This diff allows us to pinpoint where the definition key is directly.

Reviewed By: SamChou19815

Differential Revision: D76300347

fbshipit-source-id: 26168e9975666eb1d5ba142680cb25328abd3170