Add exception handler context to `identifier_at()`
Summary:
This diff enhances `identifier_at()` and teach it to recognize cases where the user cmd+click on the exception names declared in an `except` branch.

Previously, we were relying whether definition_at() can find the definition key of the function/class in the scope stack, which requires a walk of the entire scope. This diff allows us to pinpoint where the definition key is directly.

Reviewed By: SamChou19815

Differential Revision: D76358461

fbshipit-source-id: f8f75d8a8452d25e30cbf25e40a3331182f40f49