Avoid using `definition_at()` to resolve local variable names
Summary:
This diff changes the gotodef handling logic of IdentifierContext::Expr (introduced in D76174532). Previously, for a local variable definition, we were relying whether `definition_at()` can find the definition key of the variable to determine if the identifier points to a definition or usage, which not only requires a walk of the entire module scope, but is also unreliable because scopes doesn't track all definitions (see diff summary in D76289459).

In this diff we remove that dependency to `definition_at()`, and uses the expression context from the AST directly to disambiguate between defs and uses. Concretely, if expr context is `Store`, we know the identifier appears as an lvalue, so that must be a def; otherwise, the identifier must be a use. One we differentiated between defs and uses, the rest of the logic goes as before.

With this change we should now be able to do gotodefs on all shadowed defs. Right now, I'm routing all those defs into the same location which is what happened

Reviewed By: SamChou19815

Differential Revision: D76300345

fbshipit-source-id: 0989b6828a829aac21cacf6f65ac3b8f5f4ebc43