Avoid using `definition_at()` to resolve imported names
Summary: This diff rewrites the gotodef handling logic of `IdentifierContext::ImportedName` (introduced in D76289460). Previously, we were relying the fact that `definition_at()` can find the definition key of the imported name, which requires a walk of the entire module scope. In this diff we remove that dependency and uses the identifier context itself to pinpoint the key directly. Once we get the key, the rest of the logic follows exactly what happened before.

Reviewed By: SamChou19815

Differential Revision: D76300346

fbshipit-source-id: 788b3d503d0ca0035b1b125f92601ac21620c65f