Add type parameter context to `identifier_at()`
Summary:
This diff enhances `identifier_at()` and teach it to recognize cases where the user cmd+click on PEP-695-style type parameters.

Previously, we were relying whether definition_at() can find the definition key of the function/class in the scope stack, which requires a walk of the entire scope. This diff allows us to pinpoint where the definition key is directly.

Reviewed By: SamChou19815

Differential Revision: D76358459

fbshipit-source-id: de014b8ce9258f0c940e2be6b9bb6b3b345fc6ea