Extract the logic of handling jump-to-definition-itself into a separate function
Summary:
The pre-existing gotodef logic has 2 steps:
- Given a location `L`, first invoke `definition_at()` on `L` to see if `L` points to any definition in the code. If yes, return `L` immediately since we are trying to goto-def on a def itself.
- If `L` does not point to a definition, invoke `identifier_at()` on `L` obtaining the identifier and the surrounding context, then resolve the definition depending on what the context is.

Problem with this approach is that `definition_at()` is not a reliable source to find definitions: it uses scope stack for definition searching, but we know that scope stack only contains the first syntactical definition of a given name (in the static scope) and the last control-flow-sensitive definition of that same name (in the flow scope). All definitions in-between are 100% dropped, which means that the approach cannot really handle certain gotodef cases.

This stack aims at solving the problem by skipping the first step, and instead use the identifier and its context to construct definition keys for table lookup. Since the binding table does contain all intermediate definitions for any given name, this approach won't suffer from the same problem. This approach also has the additional benefit that it can now find identifiers that we cannot find previously (e.g. kwargs ids), paving the way for implementing kwarg-gotodef.

To get to a point where we could avoid using `definition_at()` altogether, the first step is to refactor pre-existing code such that the order of the two steps are swapped: we now first look up the identifier+context, and fallback to step1 only if we don't really know how to handle the identifier+context combination yet. This means that any changes we do in the identifier+context won't be "shadowed" by step1 (see my inline comments in D76289460 for an example of when this may happen) and will be immediately visible&testable. In subsequent diffs, our job now becomes to reduce the usage of `definition_at()` in the gotodef code and gradually eliminate its usage.

Reviewed By: SamChou19815

Differential Revision: D76289459

fbshipit-source-id: a71eb4baf2692d00e4306ad12e513538a7439328