improve Final/ReadOnly attrs behavior
Summary:
attributes that are declared `Final[]` or `ReadOnly[]` & are initialized on the class are now marked as read-only, preventing assignment and deletion.

this does not apply to instance-only attrs or Final attrs declared at the class level but initialized elsewhere.

marking those attrs as read-only today would raise an error if they are set in the constructor.

there isn't a good mechanism to track whether an attribute has been initialized or not, so I was thinking of allowing constructors to bypass the read-only restrictions.

for now, we only add the restriction in cases we are sure the attribute has been initialized

Reviewed By: rchen152

Differential Revision: D71062153

fbshipit-source-id: 0a1216fa463d605d4d3d0a0961e98dd34f9b226d