Add keyword argument context to `identifier_at()`
Summary:
This diff enhances `identifier_at()` and teach it to recognize cases where the user cmd+click on the name of a keyword argument, e.g. `x` in `foo(x=1)`. This is one of the most frequently appeared case in our offline experiments where Pyrefly doesn't handle well today.

The diff only adds the recognition logic and not any real handling -- that would the job for the next 2 diffs in the stack.

Reviewed By: SamChou19815

Differential Revision: D76375055

fbshipit-source-id: 97562b1a7c3757c1d6bd6dbde1243521e80ca82e